Add device cache
This ensures that reconnected devices will reuse previously created
instances if they are still referenced outside the hook.

This way if someone holds onto a reference of a disconnected device that
reference will receive all events in the event that the device is
reconnected. Previously this device would be useless and a reference to
the new instance would've been required.